+++ cricbuzz-microservices/Gateway/src/main/java/com/cricbuzz/filter/ErrorFilter.java
@@ -1,0 +1,29 @@
+package com.cricbuzz.filter;
+
+import com.netflix.zuul.ZuulFilter;
+
+public class ErrorFilter extends ZuulFilter {
+
+	@Override
+	public String filterType() {
+		return "error";
+	}
+
+	@Override
+	public int filterOrder() {
+		return 0;
+	}
+
+	@Override
+	public boolean shouldFilter() {
+		return true;
+	}
+
+	@Override
+	public Object run() {
+		System.out.println("Using Route Filter");
+
+		return null;
+	}
+
+}
