+++ cricbuzz-microservices/Gateway/src/main/java/com/cricbuzz/filter/RouteFilter.java
@@ -1,0 +1,29 @@
+package com.cricbuzz.filter;
+
+import com.netflix.zuul.ZuulFilter;
+
+public class RouteFilter extends ZuulFilter {
+
+	@Override
+	public String filterType() {
+		return "route";
+	}
+
+	@Override
+	public int filterOrder() {
+		return 0;
+	}
+
+	@Override
+	public boolean shouldFilter() {
+		return true;
+	}
+
+	@Override
+	public Object run() {
+		System.out.println("Using Route Filter");
+
+		return null;
+	}
+
+}
