+++ cricbuzz-microservices/Gateway/src/main/java/com/cricbuzz/filter/SimpleFilter.java
@@ -1,0 +1,29 @@
+package com.cricbuzz.filter;
+
+import com.netflix.zuul.ZuulFilter;
+import com.netflix.zuul.context.RequestContext;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import javax.servlet.http.HttpServletRequest;
+public class SimpleFilter extends ZuulFilter {
+   private static Logger log = LoggerFactory.getLogger(SimpleFilter.class);
+   @Override
+   public String filterType() {
+       return "pre";
+   }
+   @Override
+   public int filterOrder() {
+       return 1;
+   }
+   @Override
+   public boolean shouldFilter() {
+       return true;
+   }
+   @Override
+   public Object run() {
+       RequestContext ctx = RequestContext.getCurrentContext();
+       HttpServletRequest request = ctx.getRequest();
+       log.info(String.format("%s request to %s", request.getMethod(), request.getRequestURL().toString()));
+       return null;
+   }
+}
