--- cricbuzz-microservices/Gateway/src/main/java/com/cricbuzz/filters/PostFilter.java
@@ -1,29 +1,0 @@
-package com.cricbuzz.filters;
-
-import com.netflix.zuul.ZuulFilter;
-
-public class PostFilter extends ZuulFilter {
-
-  @Override
-  public String filterType() {
-    return "post";
-  }
-
-  @Override
-  public int filterOrder() {
-    return 1;
-  }
-
-  @Override
-  public boolean shouldFilter() {
-    return true;
-  }
-
-  @Override
-  public Object run() {
-   System.out.println("Inside Response Filter");
-
-    return null;
-  }
-
-}
