--- cricbuzz-microservices/Gateway/src/main/java/com/cricbuzz/filters/RouteFilter.java
@@ -1,29 +1,0 @@
-package com.cricbuzz.filters;
-
-import com.netflix.zuul.ZuulFilter;
-
-public class RouteFilter extends ZuulFilter {
-
-  @Override
-  public String filterType() {
-    return "route";
-  }
-
-  @Override
-  public int filterOrder() {
-    return 1;
-  }
-
-  @Override
-  public boolean shouldFilter() {
-    return true;
-  }
-
-  @Override
-  public Object run() {
-   System.out.println("Inside Route Filter");
-
-    return null;
-  }
-
-}
