+++ cricbuzz-microservices/Gateway/src/main/resources/application.properties
@@ -1,11 +1,10 @@
-eureka.client.serviceUrl.defaultZone:http://localhost:8761/eureka
+spring.application.name=Gateway
+eureka.client.eureka-server-port=8761
+server.port=1234
 security.basic.enable: false   
 management.security.enabled: false
-zuul.routes.scoreboard.url=http://localhost:2000
-zuul.routes.MatchService.url=http://localhost:3000
+zuul.host.socket-timeout-millis=30000
+zuul.routes.ScoreService.url=http://localhost:2000
 zuul.routes.PlayerService.url=http://localhost:4000
 zuul.routes.RankingService.url=http://localhost:5000
-eureka.client.eureka-server-port=8761
-spring.application.name=Gateway
-ribbon.eureka.enabled=false
-server.port=1234
+zuul.routes.MatchService.url=http://localhost:3000
