+++ cricbuzz-microservices/Gateway/src/test/java/com/cricbuzz/GatewayApplicationTests.java
@@ -1,4 +1,4 @@
-package com.example.springbootzuulgatwayproxy;
+package com.cricbuzz;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -7,10 +7,11 @@ import org.springframework.test.context.junit4.SpringRunner;
 
 @RunWith(SpringRunner.class)
 @SpringBootTest
-public class SpringBootZuulgatwayproxyApplicationTests {
+public class GatewayApplicationTests {
 
 	@Test
 	public void contextLoads() {
 	}
 
 }
+
