+++ cricbuzz-microservices/PlayersService/pom.xml
@@ -64,11 +64,11 @@
 </dependency>
 		
 <!-- https://mvnrepository.com/artifact/org.springframework.boot/spring-boot-starter-security -->
-	 <dependency>
+	<!--  <dependency>
 			<groupId>org.springframework.boot</groupId>
 			<artifactId>spring-boot-starter-security</artifactId>
 
-		</dependency>  
+		</dependency>   -->
 	</dependencies>
 
 	<build>
