+++ cricbuzz-microservices/live-scoresService/src/main/resources/application.properties
@@ -1,6 +1,7 @@
 server.port=2000
-spring.application.name=scoreboard
-
+spring.application.name=ScoreService
+spring.cloud.config.discovery.serviceid=ScoreService
+spring.cloud.config.discovery.enabled=true
 spring.datamongodb.host=localhost
 spring.data.mongodb.port=27017
 spring.data.mongodb.database=livescores
